+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/visitor.dart
@@ -1,0 +1,17 @@
+part of 'prefer_single_widget_per_file.dart';
+
+class _Visitor extends SimpleAstVisitor<void> {
+  final _nodes = <ClassDeclaration>[];
+
+  Iterable<ClassDeclaration> get nodes => _nodes;
+
+  @override
+  void visitClassDeclaration(ClassDeclaration node) {
+    super.visitClassDeclaration(node);
+
+    final classType = node.extendsClause?.superclass.type;
+    if (isWidgetOrSubclass(classType)) {
+      _nodes.add(node);
+    }
+  }
+}