+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/correct_example.dart
@@ -1,0 +1,8 @@
+import 'flutter_defines.dart';
+
+class someWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}