+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/incorrect_example.dart
@@ -1,0 +1,37 @@
+import 'flutter_defines.dart';
+
+class someWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}
+
+// LINT
+class someOtherWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}
+
+// LINT
+class _someOtherWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}
+
+// LINT
+class someStatefulWidget extends StatefulWidget {
+  @override
+  _someStatefulWidgetState createState() => _someStatefulWidgetState();
+}
+
+class _someStatefulWidgetState extends State<InspirationCard> {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}